+++ offchain/crates/solana-cli/src/command/passport/mod.rs
@@ -4,10 +4,27 @@ use doublezero_passport::state::{AccessRequest, ProgramConfig};
 use doublezero_solana_client_tools::rpc::SolanaConnection;
 use solana_sdk::pubkey::Pubkey;
 
+mod access_validation;
 pub mod fetch;
 pub mod find_validator;
 pub mod prepare_access;
 pub mod request_access;
+
+#[derive(Debug, Args, Clone)]
+pub struct SharedAccessArgs {
+    /// The DoubleZero service key to request access from
+    #[arg(long)]
+    pub doublezero_address: Pubkey,
+    /// The validator's node ID (identity pubkey)
+    #[arg(long, value_name = "PUBKEY")]
+    pub primary_validator_id: Pubkey,
+    /// Optional backup validator IDs (identity pubkeys)
+    #[arg(long, value_name = "PUBKEY,PUBKEY,PUBKEY", value_delimiter = ',')]
+    pub backup_validator_ids: Vec<Pubkey>,
+    /// Number of previous epochs to check when evaluating the leader schedule (defaults to ENV_PREVIOUS_LEADER_EPOCHS)
+    #[arg(long, hide = true)]
+    pub leader_schedule_epochs: Option<u8>,
+}
 
 #[derive(Debug, Args)]
 pub struct PassportCommand {